chore: define webpack common objects
