+++ .webpack/ts-loader.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  test: /\.ts$/,
+  loader: 'awesome-typescript-loader'
+};
