+++ .webpack/worker-loader.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  test: /\.worker\.js$/,
+  use: {
+    loader: 'worker-loader',
+    options: {
+      inline: true,
+      fallback: false,
+    }
+  }
+};
