+++ webpack.config.js
@@ -4,43 +4,17 @@ const merge = require("webpack-merge");
 const TerserPlugin = require('terser-webpack-plugin');
 const TsConfigPathsPlugin = require('tsconfig-paths-webpack-plugin');
 
+const TSLoader = require('./.webpack/ts-loader.config');
+const WorkerLoader = require('./.webpack/worker-loader.config');
+
 const PATH = {
   src: path.join(__dirname, './source'),
   build: path.join(__dirname, './dist'),
   config: path.join(__dirname, './tsconfig.json'),
 }
 
-const production = {
-  mode: 'production',
-  devtool: 'source-map',
-  optimization: {
-    minimizer: [new TerserPlugin()]
-  },
-  output: {
-    path: PATH.build,
-    filename: '[name].min.js',
-    libraryTarget: 'umd',
-    globalObject: `(typeof self !== 'undefined' ? self : this)`,
-  },
-};
-
-const debug = {
-  mode: 'development',
-  output: {
-    path: PATH.build,
-    filename: '[name].js',
-    libraryTarget: 'umd',
-    globalObject: `(typeof self !== 'undefined' ? self : this)`,
-  },
-  optimization: {
-    minimize: false,
-  },
-  plugins: [
-    new webpack.LoaderOptionsPlugin({
-      debug: true
-    })
-  ]
-};
+const production = require('./webpack.production');
+const debug = require('./webpack.debug');
 
 const base = {
   entry: {
@@ -50,22 +24,13 @@ const base = {
     'imgstry.kernel': `${PATH.src}/kernel/index.ts`,
     'imgstry': `${PATH.src}/index.ts`,
   },
+  output: {
+    path: PATH.build,
+  },
   module: {
     rules: [
-      {
-        test: /\.ts$/,
-        loader: 'awesome-typescript-loader'
-      },
-      {
-        test: /\.worker\.js$/,
-        use: {
-          loader: 'worker-loader',
-          options: {
-            inline: true,
-            fallback: false,
-          }
-        }
-      },
+      TSLoader,
+      WorkerLoader,
     ]
   },
   resolve: {
